refactor(stm): extract concatenation proof from aggregate signature
And create a new 'proof' sub-module where it is moved.
