+++ mithril-stm/src/aggregate_signature/proof/mod.rs
@@ -1,0 +1,3 @@
+mod concatenation;
+
+pub use concatenation::*;
